Refactor ActionTypes and ActionDetails into a new file
Becuase of a problem on runtime in typescript the enum was undefined. therfore extracting it into a new file would solve the problem
